+++ src/Support/Contracts/PreSaveArgResolver.php
@@ -2,4 +2,8 @@
 
 namespace Nuwave\Lighthouse\Support\Contracts;
 
+/**
+ * Resolvers implementing this interface are invoked before $model->save(),
+ * allowing them to set foreign keys on the parent model (e.g. BelongsTo).
+ */
 interface PreSaveArgResolver extends ArgResolver {}